fix test compile
